requirements.txt
# Core ML/AI
ultralytics==8.0.196
torch>=1.13.0
torchvision>=0.14.0
opencv-python==4.8.1.78
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3

# Web Framework
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Processing
roboflow==1.1.9
albumentations==1.3.1
matplotlib==3.7.2
seaborn==0.12.2

# Dataset Collection
kaggle==1.5.16
zipfile36==0.1.3

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0

# Optional: GPU support
# torch-audio  # Uncomment if needed
# torchtext    # Uncomment if needed